Move tests from test_dnatracing_methods to relevant new tracing files
