+++ tests/tracing/test_dnatracing_methods.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import numpy as np
 import pytest
 
-from topostats.tracing.dnatracing import crop_array, dnaTrace, round_splined_traces
+from topostats.tracing.dnatracing import dnaTrace, round_splined_traces
 from topostats.tracing.tracingfuncs import reorderTrace
 
 # This is required because of the inheritance used throughout
@@ -39,371 +39,6 @@ def dnatrace_spline(dnatrace: dnaTrace) -> dnaTrace:
     dnatrace.pixel_to_nm_scaling = 1.0
     dnatrace.n_grain = 1
     return dnatrace
-
-
-GRAINS = {}
-GRAINS["vertical"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["horizontal"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 1, 1, 1, 1, 1, 1, 1, 1, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["diagonal1"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 0, 0, 0, 0, 0, 1, 1, 0],
-        [0, 0, 0, 0, 0, 0, 1, 0, 0],
-        [0, 0, 0, 0, 1, 1, 1, 0, 0],
-        [0, 0, 0, 1, 1, 0, 0, 0, 0],
-        [0, 0, 1, 1, 0, 0, 0, 0, 0],
-        [0, 1, 1, 0, 0, 0, 0, 0, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["diagonal2"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0],
-        [0, 0, 0, 0, 0, 0, 1, 1, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 0],
-        [0, 0, 0, 1, 1, 0, 0, 0, 0, 0, 0],
-        [0, 0, 1, 1, 0, 0, 0, 0, 0, 0, 0],
-        [0, 1, 1, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["diagonal3"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 0, 0, 0, 0, 0, 1, 0, 0],
-        [0, 0, 0, 0, 0, 0, 1, 0, 0],
-        [0, 0, 0, 0, 1, 1, 1, 0, 0],
-        [0, 0, 0, 1, 1, 0, 0, 0, 0],
-        [0, 0, 1, 1, 0, 0, 0, 0, 0],
-        [0, 0, 1, 0, 0, 0, 0, 0, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["circle"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0],
-        [0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 0],
-        [0, 1, 1, 0, 0, 0, 0, 0, 1, 1, 0],
-        [0, 0, 1, 1, 1, 1, 1, 1, 1, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["blob"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0],
-        [0, 1, 1, 1, 0],
-        [0, 1, 1, 1, 0],
-        [0, 1, 1, 1, 0],
-        [0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["cross"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 1, 0, 0, 0],
-        [0, 0, 0, 1, 0, 0, 0],
-        [0, 1, 1, 1, 1, 1, 0],
-        [0, 0, 0, 1, 0, 0, 0],
-        [0, 0, 0, 1, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["single_L"] = np.asarray(
-    [
-        [0, 0, 0, 0],
-        [0, 0, 1, 0],
-        [0, 0, 1, 0],
-        [0, 0, 1, 0],
-        [0, 1, 1, 0],
-        [0, 0, 0, 0],
-    ]
-)
-GRAINS["double_L"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0],
-        [0, 0, 0, 1, 0],
-        [0, 0, 1, 1, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 1, 1, 0, 0],
-        [0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["diagonal_end_single_L"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0],
-        [0, 0, 0, 1, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 1, 1, 0, 0],
-        [0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["diagonal_end_straight"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0],
-        [0, 0, 0, 1, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 1, 0, 0],
-        [0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["figure8"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 1, 1, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 1, 1, 1, 1, 1, 1, 1, 0],
-        [0, 1, 0, 0, 1, 0, 0, 0, 0],
-        [0, 1, 0, 0, 1, 0, 0, 0, 0],
-        [0, 1, 1, 1, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["three_ends"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 1, 1, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 1, 1, 1, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-GRAINS["six_ends"] = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 1, 1, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 1, 1, 1, 1, 0, 0, 0, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 0, 1, 1, 1, 1, 1, 1, 0],
-        [0, 0, 0, 0, 1, 0, 0, 1, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-
-
-@pytest.mark.parametrize(
-    ("grain", "mol_is_circular"),
-    [
-        pytest.param(GRAINS["vertical"], False, id="vertical"),
-        pytest.param(GRAINS["horizontal"], False, id="horizontal"),
-        pytest.param(GRAINS["diagonal1"], True, id="diagonal1"),  # This is wrong, this IS a linear molecule
-        pytest.param(GRAINS["diagonal2"], False, id="diagonal2"),
-        pytest.param(GRAINS["diagonal3"], False, id="diagonal3"),
-        pytest.param(GRAINS["circle"], True, id="circle"),
-        pytest.param(GRAINS["blob"], True, id="blob"),
-        pytest.param(GRAINS["cross"], False, id="cross"),
-        pytest.param(GRAINS["single_L"], False, id="singl_L"),
-        pytest.param(GRAINS["double_L"], True, id="double_L"),  # This is wrong, this IS a linear molecule
-        pytest.param(GRAINS["diagonal_end_single_L"], False, id="diagonal_end_single_L"),
-        pytest.param(GRAINS["diagonal_end_straight"], False, id="diagonal_end_straight"),
-        pytest.param(GRAINS["figure8"], True, id="figure8"),
-        pytest.param(GRAINS["three_ends"], False, id="three_ends"),
-        pytest.param(GRAINS["six_ends"], False, id="six_ends"),
-    ],
-)
-def test_linear_or_circular(dnatrace, grain: np.ndarray, mol_is_circular: bool) -> None:
-    """Test the linear_or_circular method with a range of different structures."""
-    linear_coordinates = np.argwhere(grain == 1)
-    dnatrace.linear_or_circular(linear_coordinates)
-    assert dnatrace.mol_is_circular == mol_is_circular
-
-
-TEST_LABELLED = np.asarray(
-    [
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 1, 1, 1, 1, 1, 1, 0, 0, 2, 2, 2, 2, 2, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 2, 0, 0, 0, 2, 0],
-        [0, 3, 3, 3, 3, 3, 3, 0, 0, 2, 0, 0, 0, 2, 0],
-        [0, 0, 0, 0, 0, 0, 3, 0, 0, 2, 0, 0, 0, 2, 0],
-        [0, 0, 0, 0, 0, 0, 3, 0, 0, 2, 2, 2, 2, 2, 0],
-        [0, 0, 0, 0, 0, 0, 3, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 0, 0, 0, 0, 0, 3, 0, 4, 4, 4, 4, 4, 4, 0],
-        [0, 0, 0, 0, 0, 0, 3, 0, 4, 4, 4, 4, 4, 4, 0],
-        [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-        [0, 5, 5, 5, 5, 6, 6, 6, 6, 0, 0, 6, 0, 0, 0],
-        [0, 5, 5, 0, 0, 6, 0, 0, 6, 0, 0, 6, 0, 0, 0],
-        [0, 5, 5, 5, 5, 6, 0, 0, 6, 6, 6, 6, 6, 6, 0],
-        [0, 0, 0, 5, 5, 6, 6, 6, 6, 0, 0, 6, 0, 0, 0],
-        [0, 5, 5, 5, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-    ]
-)
-
-
-@pytest.mark.parametrize(
-    ("bounding_box", "target", "pad_width"),
-    [
-        pytest.param(
-            (1, 1, 2, 7),
-            np.asarray(
-                [
-                    [1, 1, 1, 1, 1, 1],
-                ]
-            ),
-            0,
-            id="Zero padding",
-        ),
-        pytest.param(
-            (1, 1, 2, 7),
-            np.asarray(
-                [
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 1, 1, 1, 1, 1, 1, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0],
-                ]
-            ),
-            1,
-            id="Single pixel padding",
-        ),
-        pytest.param(
-            (1, 1, 2, 7),
-            np.asarray(
-                [
-                    [0, 0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 1, 1, 1, 1, 1, 1, 0, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0, 0],
-                    [0, 3, 3, 3, 3, 3, 3, 0, 0],
-                ]
-            ),
-            2,
-            id="Two pixel padding",
-        ),
-        pytest.param(
-            (1, 9, 6, 14),
-            np.asarray(
-                [
-                    [2, 2, 2, 2, 2],
-                    [2, 0, 0, 0, 2],
-                    [2, 0, 0, 0, 2],
-                    [2, 0, 0, 0, 2],
-                    [2, 2, 2, 2, 2],
-                ]
-            ),
-            0,
-            id="Ring with zero padding",
-        ),
-        pytest.param(
-            (3, 1, 9, 7),
-            np.asarray(
-                [
-                    [3, 3, 3, 3, 3, 3],
-                    [0, 0, 0, 0, 0, 3],
-                    [0, 0, 0, 0, 0, 3],
-                    [0, 0, 0, 0, 0, 3],
-                    [0, 0, 0, 0, 0, 3],
-                    [0, 0, 0, 0, 0, 3],
-                ]
-            ),
-            0,
-            id="L with zero padding",
-        ),
-        pytest.param(
-            (7, 8, 9, 14),
-            np.asarray(
-                [
-                    [4, 4, 4, 4, 4, 4],
-                    [4, 4, 4, 4, 4, 4],
-                ]
-            ),
-            0,
-            id="solid with zero padding",
-        ),
-        pytest.param(
-            (10, 1, 15, 5),
-            np.asarray(
-                [
-                    [5, 5, 5, 5],
-                    [5, 5, 0, 0],
-                    [5, 5, 5, 5],
-                    [0, 0, 5, 5],
-                    [5, 5, 5, 5],
-                ]
-            ),
-            0,
-            id="small area with zero padding",
-        ),
-        pytest.param(
-            (10, 5, 14, 14),
-            np.asarray(
-                [
-                    [6, 6, 6, 6, 0, 0, 6, 0, 0],
-                    [6, 0, 0, 6, 0, 0, 6, 0, 0],
-                    [6, 0, 0, 6, 6, 6, 6, 6, 6],
-                    [6, 6, 6, 6, 0, 0, 6, 0, 0],
-                ]
-            ),
-            0,
-            id="larger area with zero pixel padding",
-        ),
-        pytest.param(
-            (10, 5, 14, 14),
-            np.asarray(
-                [
-                    [0, 0, 0, 3, 0, 4, 4, 4, 4, 4, 4, 0],
-                    [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                    [5, 5, 6, 6, 6, 6, 0, 0, 6, 0, 0, 0],
-                    [0, 0, 6, 0, 0, 6, 0, 0, 6, 0, 0, 0],
-                    [5, 5, 6, 0, 0, 6, 6, 6, 6, 6, 6, 0],
-                    [5, 5, 6, 6, 6, 6, 0, 0, 6, 0, 0, 0],
-                    [5, 5, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0],
-                ]
-            ),
-            2,
-            id="larger area with two pixel padding",
-        ),
-    ],
-)
-def test_crop_array(bounding_box: tuple, target: np.array, pad_width: int) -> None:
-    """Test the cropping of images."""
-    cropped = crop_array(TEST_LABELLED, bounding_box, pad_width)
-    np.testing.assert_array_equal(cropped, target)
 
 
 @pytest.mark.parametrize(
@@ -770,24 +405,3 @@ def test_round_splined_traces():
     expected_result = {"0": np.array([[1, 2], [3, 4]]), "1": None, "2": np.array([[6, 7], [8, 9]])}
     result = round_splined_traces(splined_traces)
     np.testing.assert_equal(result, expected_result)
-
-
-@pytest.mark.parametrize(
-    ("tuple_list", "expected_result"),
-    [
-        (
-            [(1, 2, 3), (1, 2, 3), (1, 2, 3), (1, 2, 3)],
-            [(1, 2, 3)],
-        ),
-        (
-            [(1, 2, 3), (1, 2, 3), (4, 5, 6), (4, 5, 6), (7, 8, 9), (10, 11, 12), (10, 11, 12)],
-            [(1, 2, 3), (4, 5, 6), (7, 8, 9), (10, 11, 12)],
-        ),
-        ([np.array((1, 2, 3)), np.array((1, 2, 3)), np.array((1, 2, 3)), np.array((1, 2, 3))], [(1, 2, 3)]),
-    ],
-)
-def test_remove_duplicate_consecutive_tuples(tuple_list: list[tuple], expected_result: list[tuple]) -> None:
-    """Test the remove_duplicate_consecutive_tuples function of dnatracing.py."""
-    result = dnaTrace.remove_duplicate_consecutive_tuples(tuple_list)
-
-    np.testing.assert_array_equal(result, expected_result)
